requirements.txt
seleniumbase==1.37.2
parameterized==0.7.4
selenium==3.141.0
pytest==5.3.5
PyYAML==5.3.1
